// app/_lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
      bookings: {
        Row: {
          cabin_id: number | null;
          cabin_price: number | null;
          created_at: string;
          end_date: string | null;
          extra_price: number | null;
          guest_id: number | null;
          has_breakfast: boolean | null;
          id: number;
          is_paid: boolean | null;
          num_guests: number | null;
          num_nights: number | null;
          observations: string | null;
          start_date: string | null;
          status: string | null;
          total_price: number | null;
        };
        Insert: {
          cabin_id?: number | null;
          cabin_price?: number | null;
          created_at?: string;
          end_date?: string | null;
          extra_price?: number | null;
          guest_id?: number | null;
          has_breakfast?: boolean | null;
          id?: number;
          is_paid?: boolean | null;
          num_guests?: number | null;
          num_nights?: number | null;
          observations?: string | null;
          start_date?: string | null;
          status?: string | null;
          total_price?: number | null;
        };
        Update: {
          cabin_id?: number | null;
          cabin_price?: number | null;
          created_at?: string;
          end_date?: string | null;
          extra_price?: number | null;
          guest_id?: number | null;
          has_breakfast?: boolean | null;
          id?: number;
          is_paid?: boolean | null;
          num_guests?: number | null;
          num_nights?: number | null;
          observations?: string | null;
          start_date?: string | null;
          status?: string | null;
          total_price?: number | null;
        };
        Relationships: [
          {
            foreignKeyName: "bookings_cabin_id_fkey";
            columns: ["cabin_id"];
            isOneToOne: false;
            referencedRelation: "cabins";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "bookings_guest_id_fkey";
            columns: ["guest_id"];
            isOneToOne: false;
            referencedRelation: "guests";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      guests: {
        Row: {
          country_flag: string | null;
          created_at: string;
          email: string | null;
          full_name: string | null;
          id: number;
          national_id: string | null;
          nationality: string | null;
        };
        Insert: {
          country_flag?: string | null;
          created_at?: string;
          email?: string | null;
          full_name?: string | null;
          id?: number;
          national_id?: string | null;
          nationality?: string | null;
        };
        Update: {
          country_flag?: string | null;
          created_at?: string;
          email?: string | null;
          full_name?: string | null;
          id?: number;
          national_id?: string | null;
          nationality?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R;
      }
      ? R
      : never
    : never;